+++ components/toggle-button-group/stories/arg-types.js
@@ -1,0 +1,8 @@
+export const argTypes = {
+	connotation: {
+		control: {
+			type: 'select',
+			options: ['primary'],
+		}
+	}
+};